implement rabbitmq messaging

+++ MA.RewardService.Application/Feature/HandleMissionProgress/HandleMissionProgressCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace MA.RewardService.Application.Feature.HandleMissionProgress;
 
-public record HandleMissionProgressCommand(int UserId, byte[] SpinResult): IRequest;
+public record HandleMissionProgressCommand(int UserId, int[] SpinResult): IRequest;
